# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Social Enterprise Directory
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /api/v1
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /swagger:
    x-swagger-pipe: swagger_raw

  /directory:
    x-swagger-router-controller: enterprise.controller
    get:
      description: Returns all social enterprises in the directory
      # call this method name on the controller
      operationId: getAllEnterprisesPublic
      parameters:
        - name: offset
          in: query
          description: Skip the first number of enterprises specified by offset value
          required: false
          type: number
          format: uint
        - name: count
          in: query
          description: Number of enterprises to return
          required: false
          type: number
          format: uint
          minimum: 1
          maximum: 25
        - name: q
          in: query
          description: Search keyword(s), each one separated by + symbol
          required: false
          type: string

      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
               $ref: "#/definitions/EnterprisePublic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /enterprise/{id}:
    x-swagger-router-controller: enterprise.controller
    get:
      description: Returns one social enterprise in the directory
      # call this method name on the controller
      operationId: getOneEnterprisePublic
      parameters:
        - name: id
          in: path
          description: ID of enterprise to fetch
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EnterprisePublic"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /enterprise/{id}/complete:
    x-swagger-router-controller: enterprise.controller
    get:
      description: Returns *all* details about one social enterprise in the directory
      # call this method name on the controller
      operationId: getOneEnterpriseComplete
      parameters:
        - name: id
          in: path
          description: ID of enterprise to fetch
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EnterpriseComplete"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /enterprise/{id}/logo:
    x-swagger-router-controller: enterprise.controller
    get:
      description: Returns logo of one social enterprise in the directory
      # call this method name on the controller
      operationId: getEnterpriseLogo
      produces:
        - image/jpeg
        - image/png
        - image/gif
        - image/svg+xml
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of enterprise to fetch logo for
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


  /enterprise:
    x-swagger-router-controller: enterprise.controller
    post:
      description: Creates an enterprise in the directory
      # call this method name on the controller
      operationId: createEnterprise
      parameters:
        -
          name: Enterprise
          in: body
          description: Enterprise to be added to the directory
          required: true
          schema:
            $ref: "#/definitions/NewEnterprise"

      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/EnterpriseComplete"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

# complex objects have schema definitions
definitions:

  EnterprisePublicWithoutID:
    required:
      - name
    properties:
      name:
        type: string
      logo:
        type: string
      description:
        type: string
      offering:
        type: string
      purposes:
        type: array
        items:
          type: string
      website:
        type: string
      addresses:
        type: array
        items:
           $ref: "#/definitions/Address"
      phones:
        type: array
        items:
           $ref: "#/definitions/PhoneNumber"
      emails:
        type: array
        items:
           $ref: "#/definitions/EmailAddress"
      faxes:
        type: array
        items:
           $ref: "#/definitions/FaxAddress"
      locations:
        type: array
        items:
           $ref: "#/definitions/Coordinates"

  # public view of an enterprise
  EnterprisePublic:
    allOf:
      - $ref: '#/definitions/EnterprisePublicWithoutID'
      - properties:
          id:
            type: string
    required:
      - id
      - name

  EnterprisePrivateFields:
    properties:
          parent_organization:
            type: string
          contact_person:
            type: array
            items:
              type: string
          annual_revenue_range:
            type: string
          clusters:
            type: array
            items:
              type: string
          segments:
            type: array
            items:
              type: string
          stage_of_development:
            type: string

  # full view of an enterprise (including private fields)
  EnterpriseComplete:
    allOf:
      - $ref: '#/definitions/EnterprisePublic' # inherit all the public fields
      - $ref: '#/definitions/EnterprisePrivateFields'

  # full view of an enterprise (including private fields)
  NewEnterprise:
    allOf:
      - $ref: '#/definitions/EnterprisePublicWithoutID' # inherit all the public fields
      - $ref: '#/definitions/EnterprisePrivateFields'


  PhoneNumber:
    required:
      - number
    properties:
      number:
        type: string
      public:
        type: boolean
        default: false
      tags:
        type: array
        items:
          type: string

  EmailAddress:
    required:
      - email
    properties:
      email:
        type: string
      public:
        type: boolean
        default: false
      tags:
        type: array
        items:
          type: string

  FaxAddress:
    required:
      - fax
    properties:
      fax:
        type: string
      public:
        type: boolean
        default: false
      tags:
        type: array
        items:
          type: string

  Address:
    required:
      - address
    properties:
      address:
        type: string
      public:
        type: boolean
        default: false
      tags:
        type: array
        items:
          type: string

  Coordinates:
    required:
      - lat
      - lon
    properties:
      lat:
        type: number
      lon:
        type: number

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string